IRIS-3881 | retired SingUpPageObject in favor of RegisterPage
