+++ src/test/java/com/wikia/webdriver/pageobjectsfactory/pageobject/auth/BaseAuthPage.java
@@ -1,5 +1,21 @@
 package com.wikia.webdriver.pageobjectsfactory.pageobject.auth;
 
 
-public class BaseAuthPage {
+import com.wikia.webdriver.common.logging.PageObjectLogging;
+import com.wikia.webdriver.pageobjectsfactory.componentobject.modalwindows.FacebookSignupModalComponentObject;
+import com.wikia.webdriver.pageobjectsfactory.pageobject.WikiBasePageObject;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.support.FindBy;
+
+public class BaseAuthPage extends WikiBasePageObject {
+
+  @FindBy(css = ".signup-provider-facebook")
+  private WebElement facebookSignUpButton;
+
+  public FacebookSignupModalComponentObject clickFacebookSignUp() {
+    wait.forElementClickable(facebookSignUpButton);
+    facebookSignUpButton.click();
+    PageObjectLogging.log("clickFacebookSignUp", "clicked on sign up with facebok button", true);
+    return new FacebookSignupModalComponentObject();
+  }
 }
