+++ src/test/java/com/wikia/webdriver/testcases/facebooktests/FacebookTests.java
@@ -4,9 +4,9 @@ import com.wikia.webdriver.common.core.XMLReader;
 import com.wikia.webdriver.common.core.api.GraphApi;
 import com.wikia.webdriver.common.templates.NewTestTemplate;
 import com.wikia.webdriver.pageobjectsfactory.componentobject.modalwindows.FacebookSignupModalComponentObject;
+import com.wikia.webdriver.pageobjectsfactory.pageobject.auth.RegisterPage;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.facebook.FacebookMainPageObject;
 import com.wikia.webdriver.pageobjectsfactory.pageobject.facebook.FacebookSettingsPageObject;
-import com.wikia.webdriver.pageobjectsfactory.pageobject.signup.SignUpPageObject;
 import org.testng.annotations.Test;
 import java.util.HashMap;
 
@@ -22,7 +22,7 @@ public class FacebookTests extends NewTestTemplate {
 
     new FacebookSettingsPageObject(driver).open();
     new FacebookMainPageObject(driver).login(test_user.get("email"), test_user.get("password"));
-    SignUpPageObject signUp = new SignUpPageObject(driver).open();
+    RegisterPage signUp = new RegisterPage(driver).open();
     FacebookSignupModalComponentObject fbModal = signUp.clickFacebookSignUp();
 
     String userName = "QA" + signUp.getTimeStamp();
